backend/requirements.txt
fastapi>=0.104
uvicorn>=0.24
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
xgboost>=2.0
joblib>=1.3
python-dotenv>=1.0
pydantic>=2.5
sqlalchemy>=2.0
pytest>=7.4
python-multipart>=0.0.6
